example: refactored example into smaller modules
