+++ examples/dorf_hero/src/ai/actions/mod.rs
@@ -1,0 +1,1 @@
+pub mod meander;
