+++ examples/dorf_hero/src/components/mod.rs
@@ -1,0 +1,7 @@
+pub use player::*;
+pub use position::*;
+pub use render::*;
+
+mod player;
+mod position;
+mod render;
