+++ examples/dorf_hero/src/components/player.rs
@@ -1,0 +1,3 @@
+#[derive(Default)]
+pub struct Player;
+
