+++ examples/dorf_hero/src/components/position.rs
@@ -1,0 +1,5 @@
+#[derive(Default, Copy, Clone, PartialEq)]
+pub struct Position {
+    pub x: i32,
+    pub y: i32,
+}
