+++ examples/dorf_hero/src/components/render.rs
@@ -1,0 +1,6 @@
+#[derive(Default)]
+pub struct Render {
+    pub sprite_index: usize,
+    pub sprite_order: usize,
+}
+
