+++ examples/dorf_hero/src/resources/mod.rs
@@ -1,0 +1,5 @@
+pub use game_state::*;
+pub use tile_sprite_handles::*;
+
+mod game_state;
+mod tile_sprite_handles;
