+++ examples/dorf_hero/src/resources/tile_sprite_handles.rs
@@ -1,0 +1,7 @@
+use bevy::prelude::HandleUntyped;
+
+#[derive(Default, Clone)]
+pub struct TileSpriteHandles {
+    pub handles: Vec<HandleUntyped>,
+    pub atlas_loaded: bool,
+}
